Update BalducciMovieLens.r
Though 'scales' library, for the graph 'comma' function is called, it was not listed under '!require'.

Also, where a strict minus '4' was called for when makeCluster is instantiated, this does not generalize for all users on all systems. Thus I added the 'core_subtraction_factor' var which can be hand tuned for better stability.